+++ emopic/src/main/java/mmm/emopic/app/domain/photo/repository/PhotoRepository.java
@@ -1,5 +1,6 @@
-package mmm.emopic.app.domain.photo;
+package mmm.emopic.app.domain.photo.repository;
 
+import mmm.emopic.app.domain.photo.Photo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PhotoRepository extends JpaRepository<Photo,Long> {
